img_preprocessing/app/requirements.txt
fastapi
uvicorn
celery
numpy
Pillow
opencv-python
redis